+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Upgrades [Xmlbuilder.js](https://www.npmjs.com/package/xmlbuilder) to v10 to [fix stringifying emoji](https://github.com/oozcitak/xmlbuilder-js/issues/147).
+
 ## 1.0.0 (Sep 2, 2017)
 - Changes the parsed tag name of the default XML namespace from `$foo` to `:foo` to better match other unknown namespaces.
 - Adds `--namespace` argument to the CLI.